+++ src/main/java/net/stlgamers/hittraxreporterapi/models/Report.java
@@ -12,7 +12,7 @@ public class Report {
     private String playerName;
     private Integer numberOfSwings;
     private Double avgExitVelocity;
-    private Integer maxExitVelocity;
+    private Double maxExitVelocity;
     private Double evStdDeviation;
     private Double avgLaunchAngle;
     private Double avgHhbLaunchAngle;
